+++ src/main/java/trade/paradex/api/dto/ParadexMarketSummaryDTO.java
@@ -3,62 +3,64 @@ package trade.paradex.api.dto;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class ParadexMarketSummaryDTO {
 
     @JsonProperty("ask")
-    private double ask;
+    private BigDecimal ask;
 
     @JsonProperty("ask_iv")
-    private double askIv;
+    private BigDecimal askIv;
 
     @JsonProperty("bid")
-    private double bid;
+    private BigDecimal bid;
 
     @JsonProperty("bid_iv")
-    private double bidIv;
+    private BigDecimal bidIv;
 
     @JsonProperty("created_at")
     private Long createdAt;
 
     @JsonProperty("delta")
-    private Double delta;
+    private BigDecimal delta;
 
     @JsonProperty("funding_rate")
-    private double fundingRate;
+    private BigDecimal fundingRate;
 
     @JsonProperty("future_funding_rate")
-    private double futureFundingRate;
+    private BigDecimal futureFundingRate;
 
     @JsonProperty("last_iv")
-    private double lastIv;
+    private BigDecimal lastIv;
 
     @JsonProperty("last_traded_price")
-    private double lastTradedPrice;
+    private BigDecimal lastTradedPrice;
 
     @JsonProperty("mark_iv")
-    private double markIv;
+    private BigDecimal markIv;
 
     @JsonProperty("mark_price")
-    private double markPrice;
+    private BigDecimal markPrice;
 
     @JsonProperty("open_interest")
-    private double openInterest;
+    private BigDecimal openInterest;
 
     @JsonProperty("price_change_rate_24h")
-    private double priceChangeRate24h;
+    private BigDecimal priceChangeRate24h;
 
     @JsonProperty("symbol")
     private String symbol;
 
     @JsonProperty("total_volume")
-    private double totalVolume;
+    private BigDecimal totalVolume;
 
     @JsonProperty("underlying_price")
-    private double underlyingPrice;
+    private BigDecimal underlyingPrice;
 
     @JsonProperty("volume_24h")
-    private double volume24h;
+    private BigDecimal volume24h;
 
 
 }
